requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
schedule==1.2.0
asyncio
